added reply to comment

+++ src/bot/handlers/echo_all.py
@@ -1,8 +1,10 @@
+import re
+
 from telebot.async_telebot import AsyncTeleBot
 from telebot.types import Message
 
 from src.bot.handlers.issue_attachments import attachments_issue_new_file
-from src.bot.handlers.issue_comments import comments_issue_new_text
+from src.bot.handlers.issue_comments import comments_issue_new_text, comments_issue_reply
 from src.bot.handlers.issue_create import (
     create_issue_confirm,
     create_issue_description,
@@ -22,7 +24,27 @@ def run(bot: AsyncTeleBot):
         user_state = await bot.get_state(message.from_user.id, message.chat.id)
 
         if user_state is None:
-            await bot.delete_message(message.chat.id, message.message_id)
+            reply_message = message.reply_to_message
+            if not reply_message:
+                await bot.delete_message(message.chat.id, message.message_id)
+            else:
+                if not reply_message.from_user.is_bot:
+                    await bot.delete_message(message.chat.id, message.message_id)
+                elif reply_message.text.startswith(
+                    "Вы были упомянуты в комментарии к задаче"
+                ) or reply_message.text.startswith("Был изменен комментарий к задаче"):
+                    issue_pattern = re.compile(r"задаче (\w+-\d+)")
+                    username_pattern = re.compile(r"\((\w+)\):")
+                    issue_match = issue_pattern.search(reply_message.text)
+                    username_match = username_pattern.search(reply_message.text)
+                    issue_key = issue_match.group(1) if issue_match else None
+                    username = username_match.group(1) if username_match else None
+
+                    if issue_key and username:
+                        comment_text = f"[~{username}]\n{message.text}"
+                        await comments_issue_reply(message.chat.id, issue_key, comment_text)
+                else:
+                    await bot.delete_message(message.chat.id, message.message_id)
             return
 
         # Получение логина для авторизации
